requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.0  # Для управления изменениями схемы БД
Flask-WTF>=1.0      # Для работы с формами
Flask-Login>=0.5    # Для аутентификации пользователей
Werkzeug>=2.0       # Для хеширования паролей (уже зависимость Flask, но укажем явно)
python-dotenv       # Для загрузки переменных окружения (опционально, но полезно)
email-validator     # Для валидации email в формах
